requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
fugle-trade==0.4.0
fugle-trade-core==0.4.1
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.4
importlib-metadata==6.0.0
keyring==23.5.0
keyrings.cryptfile==1.3.8
numpy==1.24.1
pika==1.3.1
prometheus-client==0.15.0
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.16.0
python-dotenv==0.21.0
requests==2.28.2
schedule==1.1.0
urllib3==1.26.14
websocket-client==1.2.1
zipp==3.11.0
